split MigrationSpec into focused test modules
